actor_locality_enabled should not be used to configure OutputSplitter locality hints
