+++ python/ray/data/_internal/execution/operators/map_operator.py
@@ -56,7 +56,6 @@ from ray.data.block import (
     to_stats,
 )
 from ray.data.context import DataContext
-from ray.util.scheduling_strategies import NodeAffinitySchedulingStrategy
 
 logger = logging.getLogger(__name__)
 
@@ -93,7 +92,6 @@ class MapOperator(OneToOneOperator, ABC):
         self._supports_fusion = supports_fusion
         self._ray_remote_args = _canonicalize_ray_remote_args(ray_remote_args or {})
         self._ray_remote_args_fn = ray_remote_args_fn
-        self._ray_remote_args_factory_actor_locality = None
         self._remote_args_for_metrics = copy.deepcopy(self._ray_remote_args)
 
         # Bundles block references up to the min_rows_per_bundle target.
@@ -244,30 +242,6 @@ class MapOperator(OneToOneOperator, ABC):
         else:
             self._output_queue = _UnorderedOutputQueue()
 
-        if options.locality_with_output:
-            if isinstance(options.locality_with_output, list):
-                locs = options.locality_with_output
-            else:
-                locs = [ray.get_runtime_context().get_node_id()]
-
-            class RoundRobinAssign:
-                def __init__(self, locs):
-                    self.locs = locs
-                    self.i = 0
-
-                def __call__(self, args):
-                    args = copy.deepcopy(args)
-                    args["scheduling_strategy"] = NodeAffinitySchedulingStrategy(
-                        self.locs[self.i],
-                        soft=True,
-                        _spill_on_unavailable=True,
-                    )
-                    self.i += 1
-                    self.i %= len(self.locs)
-                    return args
-
-            self._ray_remote_args_factory_actor_locality = RoundRobinAssign(locs)
-
         map_transformer = self._map_transformer
         # Apply additional block split if needed.
         if self.get_additional_split_factor() > 1:
@@ -312,39 +286,6 @@ class MapOperator(OneToOneOperator, ABC):
             # If the bundler has a full bundle, add it to the operator's task submission
             # queue
             self._add_bundled_input(bundled_input)
-
-    def _get_runtime_ray_remote_args(
-        self, input_bundle: Optional[RefBundle] = None
-    ) -> Dict[str, Any]:
-        ray_remote_args = copy.deepcopy(self._ray_remote_args)
-
-        # Override parameters from user provided remote args function.
-        if self._ray_remote_args_fn:
-            new_remote_args = self._ray_remote_args_fn()
-            for k, v in new_remote_args.items():
-                ray_remote_args[k] = v
-        # For tasks with small args, we will use SPREAD by default to optimize for
-        # compute load-balancing. For tasks with large args, we will use DEFAULT to
-        # allow the Ray locality scheduler a chance to optimize task placement.
-        if "scheduling_strategy" not in ray_remote_args:
-            ctx = self.data_context
-            if input_bundle and input_bundle.size_bytes() > ctx.large_args_threshold:
-                ray_remote_args[
-                    "scheduling_strategy"
-                ] = ctx.scheduling_strategy_large_args
-                # Takes precedence over small args case. This is to let users know
-                # when the large args case is being triggered.
-                self._remote_args_for_metrics = copy.deepcopy(ray_remote_args)
-            else:
-                ray_remote_args["scheduling_strategy"] = ctx.scheduling_strategy
-                # Only save to metrics if we haven't already done so.
-                if "scheduling_strategy" not in self._remote_args_for_metrics:
-                    self._remote_args_for_metrics = copy.deepcopy(ray_remote_args)
-        # This should take precedence over previously set scheduling strategy, as it
-        # implements actor-based locality overrides.
-        if self._ray_remote_args_factory_actor_locality:
-            return self._ray_remote_args_factory_actor_locality(ray_remote_args)
-        return ray_remote_args
 
     @abstractmethod
     def _add_bundled_input(self, refs: RefBundle):
